refactor(cram): mv non-cram-specific funcs to other modules
